updated file structure for views with models
