+++ shopyo/app.py
@@ -16,13 +16,13 @@ def create_app(config_name):
     login_manager.init_app(app)
     csrf = CSRFProtect(app)
 
-    from views.manufac import manufac_blueprint
-    from views.products import prod_blueprint
-    from views.settings_modif import settings_blueprint
-    from views.appointment import appointment_blueprint
-    from views.people import people_blueprint
-    from views.admin_modif import admin_blueprint
-    from views.login import login_blueprint
+    from views.manufacturer.manufac import manufac_blueprint
+    from views.products.products import prod_blueprint
+    from views.settings.settings_modif import settings_blueprint
+    from views.appointment.appointment import appointment_blueprint
+    from views.people.people import people_blueprint
+    from views.admin.admin_modif import admin_blueprint
+    from views.admin.login import login_blueprint
     from views.save import save_blueprint
 
     app.register_blueprint(manufac_blueprint)
