+++ shopyo/views/appointment/appointment.py
@@ -9,7 +9,7 @@ from flask_login import login_required, current_user
 from flask_marshmallow import Marshmallow
 
 from addon import db, ma
-from models import Appointments
+from views.appointment.models import Appointments
 from project_api import base_context
