+++ shopyo/views/people/models.py
@@ -1,0 +1,16 @@
+from addon import db
+
+
+class People(db.Model):
+    __tablename__ = 'people'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(100))
+    phone = db.Column(db.Integer)
+    mobile = db.Column(db.Integer)
+    email = db.Column(db.String(100))
+    facebook = db.Column(db.String(128))
+    twitter = db.Column(db.String(128))
+    linkedin = db.Column(db.String(128))
+    age = db.Column(db.Integer)
+    birthday = db.Column(db.String(100))
+    notes = db.Column(db.String(100))
