+++ shopyo/views/save.py
@@ -4,7 +4,7 @@ from flask import (
 from flask_login import login_required, current_user
 
 from addon import db, login_manager
-from models import Appointments
+from views.appointment.models import Appointments
 from project_api import base_context
 
 import requests
