+++ shopyo/views/settings/models.py
@@ -1,0 +1,7 @@
+from addon import db
+
+
+class Settings(db.Model):
+    __tablename__ = 'settings'
+    setting = db.Column(db.String(100), primary_key=True)
+    value = db.Column(db.String(100))
