+++ shopyo/views/settings/settings_modif.py
@@ -2,7 +2,7 @@ from flask import (
     Blueprint, render_template, request, redirect, url_for, jsonify
     )
 from addon import db
-from models import Settings
+from views.settings.models import Settings
 from flask_marshmallow import Marshmallow
 from flask_login import login_required, current_user
